+++ metadata.rb
@@ -1,8 +1,8 @@
-name             'apache2-liatrio'
-maintainer       'YOUR_COMPANY_NAME'
-maintainer_email 'YOUR_EMAIL'
+name             'apache2liatrio'
+maintainer       'Liatrio"
+maintainer_email 'drew@liatrio.com"
 license          'All rights reserved'
-description      'Installs/Configures apache2-liatrio'
+description      'Installs/Configures apache2liatrio'
 long_description IO.read(File.join(File.dirname(__FILE__), 'README.md'))
 version          '0.1.0'